renamed invoiceitem to proper snake case invoice_item
