+++ lib/invoice_item.rb
