+++ lib/sales_engine.rb
@@ -4,7 +4,7 @@ require 'time'
 require_relative 'fileio'
 require_relative 'customer_repository'
 require_relative 'invoice_repository'
-require_relative 'invoiceitem_repository'
+require_relative 'invoice_item_repository'
 require_relative 'item_repository'
 require_relative 'merchant_repository'
 require_relative 'transaction_repository'
